wip gemmini

+++ firmware/payloads/gemmini_params.h
@@ -16,8 +16,7 @@
 typedef int8_t elem_t;
 elem_t elem_t_max = 127;
 elem_t elem_t_min = -128;
-//typedef int32_t acc_t;
-typedef int8_t acc_t;
+typedef int32_t acc_t;
 typedef int64_t full_t;
 
 #define row_align(blocks) \
